Move manifest install to install_test
