--- test/integration/multicluster/source/testdata/check.multicluster.golden
@@ -1,24 +1,0 @@
-linkerd-multicluster
---------------------
-√ Link CRD exists
-√ Link resources are valid
-	* {{.LinkName}}
-√ remote cluster access credentials are valid
-	* {{.LinkName}}
-√ clusters share trust anchors
-	* {{.LinkName}}
-√ service mirror controller has required permissions
-	* {{.LinkName}}
-√ service mirror controllers are running
-	* {{.LinkName}}
-√ all gateway mirrors are healthy
-	* {{.LinkName}}
-√ all mirror services have endpoints
-√ all mirror services are part of a Link
-√ multicluster extension proxies are healthy
-‼ multicluster extension proxies are up-to-date
-    {{.ProxyVersionErr}}
-    see {{.HintURL}}l5d-multicluster-proxy-cp-version for hints
-√ multicluster extension proxies and cli versions match
-
-Status check results are √
