--- test/integration/multicluster/source/testdata/slow-cooker.yaml
@@ -1,36 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  labels:
-    test.linkerd.io/is-test-data-plane: "true"
-  name: multicluster-statefulset
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: slow-cooker
-  namespace: multicluster-statefulset
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: slow-cooker
-  template:
-    metadata:
-      annotations:
-        linkerd.io/inject: enabled
-      labels:
-        app: slow-cooker
-    spec:
-      containers:
-      - name: slow-cooker
-        image: buoyantio/slow_cooker:1.3.0
-        command:
-        - "/bin/sh"
-        args:
-        - "-c"
-        - |
-          sleep 15 # wait for pods to start
-          /slow_cooker/slow_cooker -metric-addr 0.0.0.0:9999 http://nginx-statefulset-0.nginx-statefulset-svc-target.multicluster-statefulset.svc.cluster.local:8080
-        ports:
-        - containerPort: 9999
